fix(tests): split e2e helper modules by crate
Keep the rebased rename-symbol snapshot helpers from main without
tripping dead-code in unrelated integration test crates.

Move the refactor-routing helpers into their own test-support
submodule and wire only the refactor snapshot binaries to import
it, while preserving the existing get-card harness module for the
get-card snapshots.
